+++ packages/manager/core/shell-client/src/hooks/useNavigation.ts
@@ -7,7 +7,7 @@ import { ShellContext } from '../ShellContext';
 export const useNavigation = () => {
   const { shell } = useContext(ShellContext);
 
-  return shell.navigation;
+  return shell?.navigation;
 };
 
 export default useNavigation;