+++ packages/manager/core/shell-client/src/hooks/useTracking.ts
@@ -7,7 +7,7 @@ import { ShellContext } from '../ShellContext';
 export const useTracking = () => {
   const { shell } = useContext(ShellContext);
 
-  return shell.tracking;
+  return shell?.tracking;
 };
 
 export default useTracking;