+++ packages/manager/core/shell-client/src/hooks/useUX.ts
@@ -5,7 +5,7 @@ import useShell from './useShell';
  */
 export const useUX = () => {
   const shell = useShell();
-  return shell.ux;
+  return shell?.ux;
 };
 
 export default useUX;