+++ packages/manager/apps/container/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.112.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-container-app@0.112.0...@ovh-ux/manager-container-app@0.112.1) (2025-03-27)
+
+**Note:** Version bump only for package @ovh-ux/manager-container-app
+
+
+
+
+
 # [0.112.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-container-app@0.111.1...@ovh-ux/manager-container-app@0.112.0) (2025-03-25)