+++ packages/manager/apps/zimbra/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-zimbra-app@0.19.13...@ovh-ux/manager-zimbra-app@0.19.14) (2025-03-27)
+
+**Note:** Version bump only for package @ovh-ux/manager-zimbra-app
+
+
+
+
+
 ## [0.19.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-zimbra-app@0.19.12...@ovh-ux/manager-zimbra-app@0.19.13) (2025-03-24)