+++ packages/manager/core/shell-client/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-react-shell-client@0.8.6...@ovh-ux/manager-react-shell-client@0.9.0) (2025-03-27)
+
+
+### Features
+
+* **shell-client:** add useHidePreloader hook ([86e89ec](https://github.com/ovh/manager/commit/86e89ec6b31a73d13b6e5d587d78289016e73bca))
+
+
+
+
+
 ## [0.8.6](https://github.com/ovh/manager/compare/@ovh-ux/manager-react-shell-client@0.8.5...@ovh-ux/manager-react-shell-client@0.8.6) (2025-02-06)
 
 **Note:** Version bump only for package @ovh-ux/manager-react-shell-client