+++ packages/manager/modules/logs-to-customer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/ovh/manager/compare/@ovh-ux/logs-to-customer@1.1.0...@ovh-ux/logs-to-customer@1.1.1) (2025-03-27)
+
+**Note:** Version bump only for package @ovh-ux/logs-to-customer
+
+
+
+
+
 # [1.1.0](https://github.com/ovh/manager/compare/@ovh-ux/logs-to-customer@1.0.0...@ovh-ux/logs-to-customer@1.1.0) (2025-03-26)